% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBVhelpR-package.R
\docType{package}
\name{EBVhelpR-package}
\alias{EBVhelpR}
\alias{EBVhelpR-package}
\title{EBVhelpR: EBV DLBCL Project Data Help in R}
\description{
Provides data loading, harmonization and plotting helpers for the EBV DLBCL project. Covers whole genome sequencing, Phenocycler, RNAscope and cosMx assays, resolving per-assay source files and joining them to EBER status metadata.
}
\author{
\strong{Maintainer}: Joseph Boyd \email{boydrj7@gmail.com}

Authors:
\itemize{
  \item Joseph Boyd \email{boydrj7@gmail.com}
}

}
\keyword{internal}
