% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{get_original_cells_per_sample_id}
\alias{get_original_cells_per_sample_id}
\title{Get Original Cell Counts Per Sample}
\usage{
get_original_cells_per_sample_id(object)
}
\arguments{
\item{object}{A \link[=CellDataStore-class]{CellDataStore} object.}
}
\value{
Data frame with \code{sample_id} and \code{n_cells_original} columns.
}
\description{
Get Original Cell Counts Per Sample
}
\examples{
\dontrun{
get_original_cells_per_sample_id(cds)
}
}
