% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{get_selected_cell_data}
\alias{get_selected_cell_data}
\title{Get Selected Cell Data}
\usage{
get_selected_cell_data(object)
}
\arguments{
\item{object}{A \link[=CellDataStore-class]{CellDataStore} object.}
}
\value{
The selected cell-level data frame.
}
\description{
Get Selected Cell Data
}
\examples{
\dontrun{
sel_df <- get_selected_cell_data(cds)
}
}
