% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{new_cell_data_store}
\alias{new_cell_data_store}
\title{Construct a CellDataStore Object}
\usage{
new_cell_data_store(cell_df)
}
\arguments{
\item{cell_df}{Cell-level data frame with a \code{sample_id} column.}
}
\value{
A validated \link[=CellDataStore-class]{CellDataStore} object.
}
\description{
Creates a \link[=CellDataStore-class]{CellDataStore} where \code{selected_df} is initialized to
\code{full_df}, and original per-sample cell counts are stored in
\code{original_cells_per_sample_id}.
}
\examples{
\dontrun{
q <- CellQuery()
cell_df <- load_query_cell_data(q)
cds <- new_cell_data_store(cell_df)
}
}
