% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs.R
\name{plot_wgs_pileup_lines}
\alias{plot_wgs_pileup_lines}
\title{Plot EBV bigwig pileup line profiles by status}
\usage{
plot_wgs_pileup_lines(pileup_df, ylim = c(0, 2000))
}
\arguments{
\item{pileup_df}{Output from \code{\link[=load_wgs_bigwig_pileup]{load_wgs_bigwig_pileup()}}.}

\item{ylim}{Numeric length-2 vector giving y limits.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot EBV bigwig pileup line profiles by status
}
