% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs.R
\name{plot_wgs_viral_enrichment}
\alias{plot_wgs_viral_enrichment}
\title{Plot WGS viral enrichment by sample}
\usage{
plot_wgs_viral_enrichment(wgs_count_summary, status_colors = NULL)
}
\arguments{
\item{wgs_count_summary}{Output from \code{\link[=load_wgs_count_summary]{load_wgs_count_summary()}}.}

\item{status_colors}{Named vector of colors keyed by \code{EBER_status}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot WGS viral enrichment by sample
}
