% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\name{CellQuery}
\alias{CellQuery}
\title{Construct a CellQueryInfo Object}
\usage{
CellQuery(assay_type = NULL)
}
\arguments{
\item{all_cell_files_df}{Data frame of all available cell files.}
}
\value{
A validated \linkS4class{CellQueryInfo} object.
}
\description{
Construct a CellQueryInfo Object
}
\examples{
CellQuery()
assay_type = EBV_ASSAY_TYPES$phenocycler
all_cell_files_df = NULL
selected_sample_ids = character(0)
tiff_paths_df = NULL
get_tiff_file_path_df = EBVhelpR:::get_tiff_file_path_df
}
