% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\docType{class}
\name{CellQueryInfo-class}
\alias{CellQueryInfo-class}
\title{Cell Query Information Container}
\description{
S4 class for storing data used during cell-query image retrieval workflows.
}
\section{Slots}{

\describe{
\item{\code{all_cell_files_df}}{Data frame of all available cell files.}

\item{\code{selected_cell_files_df}}{Data frame of cell files selected for a query.}

\item{\code{tiff_paths_df}}{Data frame mapping samples to TIFF image paths.}

\item{\code{assay_type}}{Character scalar indicating assay type (for example,
\code{"RNAScope_4plex"}).}
}}

