% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_options.R
\docType{data}
\name{EBV_CHANNELS}
\alias{EBV_CHANNELS}
\title{Channel identities in tiff files.}
\format{
A named list with three character elements:
\describe{
\item{Phenocycler}{\code{"Phenocycler"}}
\item{RNAScope_4plex}{\code{"RNAScope_4plex"}}
\item{RNAScope_3plex+IF}{\code{"RNAScope_3plex+IF"}}
}
}
\usage{
EBV_CHANNELS
}
\description{
Named list of channel names.
}
\examples{
EBV_CHANNELS$RNAScope_4plex
}
\keyword{datasets}
