% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_options.R
\docType{data}
\name{EBV_OPAL_DECODE}
\alias{EBV_OPAL_DECODE}
\title{Named by OpalDye string and value are targets}
\format{
A named list with three character elements:
\describe{
\item{RNAScope_4plex}{\code{"RNAScope_4plex"}}
\item{RNAScope_3plex+IF}{\code{"RNAScope_3plex+IF"}}
}
}
\usage{
EBV_OPAL_DECODE
}
\description{
Named list of channel names.
}
\examples{
EBV_OPAL_DECODE$RNAScope_4plex
}
\keyword{datasets}
