% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBVhelpR-package.R
\docType{package}
\name{EBVhelpR-package}
\alias{EBVhelpR}
\alias{EBVhelpR-package}
\title{EBVhelpR: EBV DLBCL Project Data Help in R}
\description{
Support for analysis of EBV DLBCL project data - WGS, Phenocycler, RNAscope, cosMx
}
\author{
\strong{Maintainer}: Joseph Boyd \email{boydrj7@gmail.com}

Authors:
\itemize{
  \item Joseph Boyd \email{boydrj7@gmail.com}
}

}
\keyword{internal}
