% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\name{filter_query_to_tiff_path_samples}
\alias{filter_query_to_tiff_path_samples}
\title{Filter Query Data To TIFF-Path Sample IDs}
\usage{
filter_query_to_tiff_path_samples(object)
}
\arguments{
\item{object}{A \linkS4class{CellQueryInfo} object.}
}
\value{
Updated \linkS4class{CellQueryInfo} object.
}
\description{
Restricts all stored data frames in a \linkS4class{CellQueryInfo} object to
sample ids present in \code{tiff_paths_df}.
}
