% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{get_full_cell_data}
\alias{get_full_cell_data}
\title{Get Full Cell Data}
\usage{
get_full_cell_data(object)
}
\arguments{
\item{object}{A \linkS4class{CellDataStore} object.}
}
\value{
The full cell-level data frame.
}
\description{
Get Full Cell Data
}
\examples{
\dontrun{
full_df <- get_full_cell_data(cds)
}
}
