% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{get_original_cell_data_dir}
\alias{get_original_cell_data_dir}
\title{Title}
\usage{
get_original_cell_data_dir()
}
\value{
Character scalar path to the discovered original cell data directory.
}
\description{
Title
}
\examples{
\dontrun{
get_original_cell_data_dir()
}
}
