% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\name{get_query_cell_files_df}
\alias{get_query_cell_files_df}
\title{Get Cell File Rows From CellQueryInfo}
\usage{
get_query_cell_files_df(object, selected_only = TRUE)
}
\arguments{
\item{object}{A \linkS4class{CellQueryInfo} object.}

\item{selected_only}{Logical; if \code{TRUE}, return only selected samples.}
}
\value{
Data frame of cell file rows.
}
\description{
Get Cell File Rows From CellQueryInfo
}
