% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\name{get_query_summary_df}
\alias{get_query_summary_df}
\title{Get Summary Rows From CellQueryInfo}
\usage{
get_query_summary_df(object, selected_only = TRUE)
}
\arguments{
\item{object}{A \linkS4class{CellQueryInfo} object.}

\item{selected_only}{Logical; if \code{TRUE}, return only selected samples.}
}
\value{
Data frame of summary rows.
}
\description{
Get Summary Rows From CellQueryInfo
}
