% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{get_tiff_file_path_df}
\alias{get_tiff_file_path_df}
\title{.get_valid_project_names = EBVhelpR:::.get_valid_project_names
.get_example_images_df = EBVhelpR:::.get_example_images_df}
\usage{
get_tiff_file_path_df()
}
\description{
.get_valid_project_names = EBVhelpR:::.get_valid_project_names
.get_example_images_df = EBVhelpR:::.get_example_images_df
}
