% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils-LAPTOP-T6S5THI3.R, R/io_utils.R
\name{get_wrangled_cell_data_dir}
\alias{get_wrangled_cell_data_dir}
\title{Get Wrangled Cell Data Directory}
\usage{
get_wrangled_cell_data_dir()

get_wrangled_cell_data_dir()
}
\value{
Character scalar path to the package wrangled cell data directory.

Character scalar path to the package wrangled cell data directory.
}
\description{
Returns the directory where processed package-ready cell data files are
written. This path is derived from \code{\link[=get_original_cell_data_dir]{get_original_cell_data_dir()}}.
}
\examples{
\dontrun{
get_wrangled_cell_data_dir()
}
\dontrun{
get_wrangled_cell_data_dir()
}
}
