% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{load_query_cell_data_store}
\alias{load_query_cell_data_store}
\title{Load Query Cell Data As CellDataStore}
\usage{
load_query_cell_data_store(object)
}
\arguments{
\item{object}{A \linkS4class{CellQueryInfo} object.}
}
\value{
A \linkS4class{CellDataStore} object.
}
\description{
Convenience wrapper around \code{\link[=load_query_cell_data]{load_query_cell_data()}} + \code{\link[=new_cell_data_store]{new_cell_data_store()}}.
}
\examples{
\dontrun{
q <- CellQuery()
cds <- load_query_cell_data_store(q)
}
}
