% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs.R
\name{load_wgs_reference_genome}
\alias{load_wgs_reference_genome}
\title{Load WGS reference ranges}
\usage{
load_wgs_reference_genome()
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object.
}
\description{
Reads the package chrSizes.txt file.
}
\details{
\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object with sequence lengths set.
}
