% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\name{new_cell_query_info}
\alias{new_cell_query_info}
\title{Construct a CellQueryInfo Object}
\usage{
new_cell_query_info(
  assay_type = NULL,
  summary_df = NULL,
  all_cell_files_df = NULL,
  selected_cell_files_df = all_cell_files_df,
  tiff_paths_df = NULL
)
}
\arguments{
\item{assay_type}{Character scalar indicating assay type.}

\item{all_cell_files_df}{Data frame of all available cell files.}

\item{selected_cell_files_df}{Data frame of selected cell files.}

\item{tiff_paths_df}{Data frame with TIFF path information.}
}
\value{
A validated \linkS4class{CellQueryInfo} object.
}
\description{
Construct a CellQueryInfo Object
}
\examples{
assay_type = EBV_ASSAY_TYPES$phenocycler
all_cell_files_df = NULL
selected_cell_files_df = all_cell_files_df
tiff_paths_df = NULL
get_tiff_file_path_df = EBVhelpR:::get_tiff_file_path_df
}
