% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{replace_full_with_selected_cell_data}
\alias{replace_full_with_selected_cell_data}
\title{Replace Full Data With Selected Data}
\usage{
replace_full_with_selected_cell_data(object)
}
\arguments{
\item{object}{A \linkS4class{CellDataStore} object.}
}
\value{
Updated \linkS4class{CellDataStore} object.
}
\description{
Commits the currently selected view by replacing \code{full_df} with
\code{selected_df}.
}
\examples{
\dontrun{
cds <- replace_full_with_selected_cell_data(cds)
}
}
