% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{restore_selected_cell_data}
\alias{restore_selected_cell_data}
\title{Restore Selected Data To Full Data}
\usage{
restore_selected_cell_data(object)
}
\arguments{
\item{object}{A \linkS4class{CellDataStore} object.}
}
\value{
Updated \linkS4class{CellDataStore} object.
}
\description{
Resets \code{selected_df} to the complete \code{full_df}.
}
\examples{
\dontrun{
cds <- restore_selected_cell_data(cds)
}
}
