% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class.R
\name{set_selected_sample_ids}
\alias{set_selected_sample_ids}
\title{Set Selected Sample IDs}
\usage{
set_selected_sample_ids(object, selected_sample_ids)
}
\arguments{
\item{object}{A \link[=CellQueryInfo-class]{CellQueryInfo} object.}

\item{selected_sample_ids}{Character vector of selected sample ids.}
}
\value{
Updated \link[=CellQueryInfo-class]{CellQueryInfo} object.
}
\description{
Set Selected Sample IDs
}
\examples{
q <- CellQuery()
q <- set_selected_sample_ids(q, head(get_query_summary_df(q)$sample_id, 2))
}
