% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data_store.R
\name{show,CellDataStore-method}
\alias{show,CellDataStore-method}
\title{Display a CellDataStore Summary}
\usage{
\S4method{show}{CellDataStore}(object)
}
\arguments{
\item{object}{A \linkS4class{CellDataStore} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Display a CellDataStore Summary
}
