% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_class-LAPTOP-T6S5THI3.R
\name{show,CellQueryInfo-method}
\alias{show,CellQueryInfo-method}
\title{Display a CellQueryInfo Summary}
\usage{
\S4method{show}{CellQueryInfo}(object)
}
\arguments{
\item{object}{A \linkS4class{CellQueryInfo} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints assay name and key counts for total versus selected samples, cells,
and available TIFF files.
}
\examples{
q <- CellQuery()
show(q)
}
